requirements.txt
PyYAML==3.11
Twisted==14.0.2
Werkzeug==0.9.6
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.18
celery==3.1.16
cffi==0.8.6
characteristic==14.2.0
cryptography==0.6.1
klein==0.2.3
kombu==3.0.23
nose==1.3.4
pyOpenSSL==0.14
pyasn1==0.1.7
pyasn1-modules==0.0.5
pycparser==2.10
pytz==2014.9
service-identity==14.0.0
six==1.8.0
treq==0.2.1
txCelery==1.0.1
wsgiref==0.1.2
zope.interface==4.1.1
